Add relationships Product/Category/Ingredient, remove jointable entities and repos
