+++ data/src/main/java/me/fontys/semester4/data/entity/Category.java
@@ -1,36 +1,46 @@
 package me.fontys.semester4.data.entity;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
+import javax.persistence.*;
+import java.io.Serial;
 import java.io.Serializable;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 @Entity
 @Table(name = "category")
 public class Category implements Serializable {
 
+    @Serial
     private static final long serialVersionUID = 1L;
 
     @Id
     @Column(name = "catid", nullable = false)
     private Long catid;
 
-    @Column(name = "parentid")
-    private Long parentid;
-
     @Column(name = "name")
     private String name;
 
-    protected Category() {
+    @ManyToOne
+    @JoinColumn(name="catid")
+    private Category parent;
 
+    @OneToMany(mappedBy = "parent")
+    private List<Category> children;
+
+    @ManyToMany(mappedBy = "categories")
+    private Set<Product> products;
+
+    public Category(Long catid, Category parent, String name) {
+        this.catid = catid;
+        this.parent = parent;
+        this.name = name;
+        products = new HashSet<>();
     }
 
-    public Category(Long catid, Long parentid, String name) {
-        this.catid = catid;
-        this.parentid = parentid;
-        this.name = name;
+    public Category() {
+
+        products = new HashSet<>();
     }
 
     public void setCatid(Long catid) {
@@ -41,27 +51,35 @@ public class Category implements Serializable {
         return catid;
     }
 
-    public void setParentid(Long parentid) {
-        this.parentid = parentid;
+    public Category getParent() {
+        return parent;
     }
 
-    public Long getParentid() {
-        return parentid;
-    }
-
-    public void setName(String name) {
-        this.name = name;
+    public void setParent(Category parent) {
+        this.parent = parent;
     }
 
     public String getName() {
         return name;
     }
 
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public List<Category> getChildren() {
+        return children;
+    }
+
+    public void setChildren(List<Category> children) {
+        this.children = children;
+    }
+
     @Override
     public String toString() {
         return "Category{" +
                 "catid=" + catid + '\'' +
-                "parentid=" + parentid + '\'' +
+                "parentid=" + parent.getCatid() + '\'' +
                 "name=" + name + '\'' +
                 '}';
     }
