+++ dominos/src/main/java/me/fontys/semester4/dominos/converter/ingredientConverter.java
@@ -1,0 +1,4 @@
+package me.fontys.semester4.dominos.converter;
+
+public class ingredientConverter {
+}
